+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.88.0 - (May 15, 2024)
+
+* Added
+  * Added examples for drill-in view under `terra-navigation-side-menu`
+
 ## 1.87.0 - (May 14, 2024)
 
 * Changed